Breakup the cmake
This PR breaks up the cmake into more manageable chunks and factors lots of things into guarantee_
methods to make sure preconditions are met only when needed. it also removes a lot of the CACHE SET
stuff we had to do for subordinate cmake.

It also adds some rudimentary documentation to help navigate the new files.

It also adds an XCode stage to the CI build; and builds the Standalone in CI on all platforms
in CPM download mode

+++ src/detail/standalone/standalone_host.cpp
@@ -122,7 +122,7 @@ void StandaloneHost::clapProcess(void *pOutput, const void *pInput, uint32_t fra
     if (mainInIdx < 0)
     {
       // TODO cleaner
-      mainInIdx = ptrIdx - 2;
+      mainInIdx = (int32_t)(ptrIdx - 2);
     }
     bufIdx++;
   }
@@ -143,7 +143,7 @@ void StandaloneHost::clapProcess(void *pOutput, const void *pInput, uint32_t fra
     if (mainOutIdx < 0)
     {
       // TODO cleaner
-      mainOutIdx = ptrIdx - 2;
+      mainOutIdx = (int32_t)(ptrIdx - 2);
     }
 
     bufIdx++;
